Modifica gestione errori per rotta multa e transito

+++ backend/src/dao/transitoDao.ts
@@ -1,20 +1,19 @@
-import { DAO } from "./daoInterface";
 import Transito, { ITransitoAttributes, ITransitoCreationAttributes } from "../models/transito";
 import { Transaction } from "sequelize";
 import { HttpErrorFactory, HttpErrorCodes } from '../utils/errorHandler';
 
 // Interfaccia TransitoDAO che estende la DAO per includere metodi specifici per Transito
-interface ITransitoDAO extends DAO<ITransitoAttributes, number> {
+/* interface ITransitoDAO extends DAO<ITransitoAttributes, number> {
     // metodi da aggiungere nel caso specifico dei transiti
-}
+} */
 
 // Classe TransitoDao che implementa l'interfaccia TransitoDAO
-class TransitoDao implements ITransitoDAO {
+class TransitoDao /* implements ITransitoDAO */ {
 
     /**
      * Funzione per ottenere tutti i transiti.
      *
-     * @returns - Una promessa che risolve con un array di transiti.
+     * @returns {Promise<Transito[]>} - Una promessa che risolve con un array di transiti.
      */
     public async getAll(): Promise<Transito[]> {
         try {
@@ -28,19 +27,14 @@ class TransitoDao implements ITransitoDAO {
      * Funzione per ottenere un transito da un ID.
      * 
      * @param id - L'ID da utilizzare per ottenere il transito.
-     * @returns - Una promessa che risolve con il transito trovato.
+     * @returns {Promise<Transito>} - Una promessa che risolve con il transito trovato.
      */
-    public async getById(id: number): Promise<Transito | null> {
-        try {
-            const transito = await Transito.findByPk(id);
-            if (!transito) {
-                throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non trovato.`);
-            } else {
-                return transito;
-            }
-        } catch (error) {
-            throw error;
-            ;
+    public async getById(id: number): Promise<Transito> {
+        const transito = await Transito.findByPk(id);
+        if (!transito) {
+            throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non trovato.`);
+        } else {
+            return transito;
         }
     }
 
@@ -48,7 +42,7 @@ class TransitoDao implements ITransitoDAO {
      * Funzione per creare un nuovo transito.
      * 
      * @param transito - L'oggetto transito da creare.
-     * @returns - Una promessa che risolve con il transito creato.
+     * @returns {Promise<Transito>} - Una promessa che risolve con il transito creato.
      */
     public async create(transito: ITransitoCreationAttributes, options?: { transaction?: Transaction }): Promise<Transito> {
         try {
@@ -63,23 +57,19 @@ class TransitoDao implements ITransitoDAO {
      * 
      * @param id - L'ID del transito da aggiornare.
      * @param transito - L'oggetto transito da aggiornare.
-     * @returns - Una promessa che risolve con il numero di righe aggiornate e l'array di transiti aggiornati.
+     * @returns {Promise<[number, Transito[]]>} - Una promessa che risolve con il numero di righe aggiornate e l'array di transiti aggiornati.
      */
     public async update(id: number, transito: ITransitoAttributes): Promise<[number, Transito[]]> {
-        try {
-            const existingTransito = await Transito.findByPk(id);
-            if (!existingTransito) {
-                throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non trovato.`);
-            }
-            const [row, updatedTransito] = await Transito.update(transito, { where: { id_transito: id }, returning: true });
-            if (row === 0) {
-                throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non aggiornato.`);
-            }
-
-            return [row, updatedTransito];
-        } catch (error) {
-            throw error;
+        const existingTransito = await Transito.findByPk(id);
+        if (!existingTransito) {
+            throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non trovato.`);
         }
+        const [row, updatedTransito] = await Transito.update(transito, { where: { id_transito: id }, returning: true });
+        if (row === 0) {
+            throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non aggiornato.`);
+        }
+
+        return [row, updatedTransito];
     }
 
     /**
@@ -87,15 +77,16 @@ class TransitoDao implements ITransitoDAO {
      * 
      * @param id - L'ID del transito da eliminare.
      * @param options - Opzioni per la transazione.
-     * @returns - Una promessa che risolve con il numero di righe eliminate.
+     * @returns - Una promessa che risolve con il numero di righe eliminate e il transito eliminato.
      */
-    public async delete(id: number, options?: { transaction?: Transaction }): Promise<number> {
+    public async delete(id: number, options?: { transaction?: Transaction }): Promise<[number, Transito]> {
+        const transito = await Transito.findByPk(id);
+        if (!transito) {
+            throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non trovato.`);
+        }
         try {
-            const deleted = await Transito.destroy({ where: { id_transito: id }, ...options });
-            if (deleted === 0) {
-                throw HttpErrorFactory.createError(HttpErrorCodes.NotFound, `Transito con ID ${id} non trovato.`);
-            }
-            return deleted;
+            const rows = await Transito.destroy({ where: { id_transito: id }, ...options });
+            return [rows,transito]
         } catch {
             throw HttpErrorFactory.createError(HttpErrorCodes.InternalServerError, `Errore nell'eliminazione del transito con ID ${id}.`);
         }
